+++ config_files/test.conf
@@ -27,7 +27,7 @@ connection {
 
 		# /upload location is a location that able file upload on the server.
 		location /upload {
-			method PUT;
+			method GET;
 			upload_enable on;
 			upload_path /tmp/uploads/;
 			;
